+++ backend/canisters/notifications_index/impl/src/lib.rs
@@ -40,10 +40,6 @@ impl RuntimeState {
         self.data.governance_principals.contains(&caller)
     }
 
-    pub fn is_caller_user_index(&self) -> bool {
-        self.env.caller() == self.data.user_index_canister_id
-    }
-
     pub fn is_caller_push_service(&self) -> bool {
         self.data.push_service_principals.contains(&self.env.caller())
     }