+++ backend/canisters/notifications_index/impl/src/updates/mod.rs
@@ -1,5 +1,4 @@
 mod add_notifications_canister;
-mod c2c_update_user_principal;
 mod push_subscription;
 mod remove_subscription;
 mod remove_subscriptions;