+++ backend/canisters/user_index/impl/src/jobs/mod.rs
@@ -1,18 +1,14 @@
 use crate::RuntimeState;
 pub mod make_pending_payments;
-pub mod notify_user_principal_updates;
 pub mod submit_message_to_modclub;
 pub mod sync_events_to_local_user_index_canisters;
-pub mod sync_legacy_user_principals;
 pub mod sync_users_to_storage_index;
 pub mod upgrade_canisters;
 
 pub(crate) fn start(state: &RuntimeState) {
     make_pending_payments::start_job_if_required(state);
-    notify_user_principal_updates::start_job_if_required(state);
     submit_message_to_modclub::start_job_if_required(state);
     sync_events_to_local_user_index_canisters::start_job_if_required(state);
-    sync_legacy_user_principals::start_job_if_required(state);
     sync_users_to_storage_index::start_job_if_required(state);
     upgrade_canisters::start_job_if_required(state);
 }