+++ backend/canisters/notifications_index/CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Fix 'out of cycles' check to use new response code ([#5503](https://github.com/open-chat-labs/open-chat/pull/5503))
 
+### Removed
+
+- Remove code to update user principals ([#5808](https://github.com/open-chat-labs/open-chat/pull/5808))
+
 ## [[2.0.1026](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1026-notifications_index)] - 2024-01-25
 
 ### Added